Segregated STRING ingest utility variables and methods into a separate string_utils.py module, to keep the main ingest file uncluttered
